+++ source/Lib/TLibEncoder/TEncSearch.cpp
@@ -1208,7 +1208,7 @@ Void TEncSearch::xIntraCodingLumaBlk(TComDataCU* pcCU,
         {
             for (UInt uiX = 0; uiX < uiWidth; uiX++)
             {
-                pReco[uiX] = ClipY(static_cast<Short>(pPred[uiX]) + pResi[uiX]);
+                pReco[uiX] = ClipY(static_cast<Short>(pPred[uiX]) + static_cast<Short>(pResi[uiX]));
                 pRecQt[uiX] = pReco[uiX];
                 pRecIPred[uiX] = pReco[uiX];
             }
